pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.43.1",
    "mcp>=1.2.0",
    "openai>=1.59.8",
    "python-dotenv>=1.0.1",
]
